waste_classification/waste_classification/requirement.txt
google-genai
functions-framework
requests